Increased speed of basic projectiles, added trail to basic projectiles for visibility, added explosive projectiles, and converted all health values from integers to floats.

+++ groupMobileGame/Assets/Scripts/EnemyBase.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EnemyBase : MonoBehaviour
 {
-    public int Health = 100;
+    public float Health = 100;
     float DamagedDelay = 1;
     float SpawnDelay1;
     float SpawnDelay2;
